requirements.txt
setuptools
nose
six
numpy
astropy
python-dateutil
matplotlib
seaborn
